MTHGNN-master_test/requirements.txt
certifi==2021.5.30
cffi 
click==8.0.4
colorama==0.4.5
cycler==0.11.0
decorator==4.4.2
h5py==2.10.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.0
kiwisolver==1.3.1
matplotlib==3.3.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
networkx==2.5.1
nltk==3.6.7
olefile 
pandas==1.1.0
Pillow 
protobuf==3.19.6
pycparser 
pyparsing==3.0.9
pyreadline==2.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==5.4.1
regex==2022.9.13
scikit-learn==0.24.2
scipy==1.5.4
six 
sklearn==0.0
tensorboardX==2.4.1
threadpoolctl==3.1.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.64.1
typing_extensions==4.1.1
wincertstore==0.2
xlrd==2.0.1
zipp==3.6.0
